requirements.txt
# Core
pandas>=2.0
numpy>=1.24
tqdm>=4.65
requests>=2.31
python-dotenv>=1.0

# Vision / image processing
opencv-python>=4.8
Pillow>=10.0

# Model providers
google-genai>=0.1.0
openai>=1.0
# OpenRouter uses the requests library directly (no SDK)

# NLP metrics
nltk>=3.8
rouge>=1.0.1
rouge-score>=0.1.2
sacrebleu>=2.3.0
